feat: Commit for USART2 example on STM32F446RE
Adds an example project demonstrating USART2 communication on the
STM32F446RE Nucleo board. The example sends a value over USART2
back to a connected terminal. This can be read by using the screen
command on Linux or a serial terminal program on other OSes.

This example will be used for future debugging and testing of the
different sensors connected to the board.

Breaking changes: None

Resolves: None